+++ www/src/markdown.js
@@ -1,13 +1,14 @@
 "use strict";
 
 const MarkdownIt = require("markdown-it");
+const {full: emoji} = require("markdown-it-emoji");
 
 const renderer = new MarkdownIt({
     breaks: true,
     html: false,
     linkify: true,
     typographer: false
-});
+}).use(emoji, {shortcuts: {}});
 
 function renderMarkdown(source) {
     return renderer.render(source || "");